daemon: Introduce DRIVER_MODE variable
Introduce a variable to declaratively switch the device to driver mode.

An intended side effect is that we now switch to driver mode earlier,
*before* applying persistence.

This fixes a problem for the BlackWidow V4 Pro, where the keyboard would
reset to 1000 Hz after driver mode, defeating the 8000 Hz applied by
persistence.

Fixes #2408